Delete / Update Bug fix

+++ webshop-backend/Repositories/IProductsRespository.cs
@@ -4,7 +4,7 @@ namespace webshop_backend.Repositories
 {
     public interface IProductsRespository
     {
-        void CreateProduct(string name, string description, int price, int quantity);
+        void CreateProduct(string name, string description, decimal price, int quantity);
         void UpdateProduct(int id, string name, string description, decimal price, int quantity);
         IList<ProductsModel> GetProducts();
         void DeleteProduct(int id);
